// packages/badanmu/src/kuaishou/lib.d.ts
/** Declaration file generated by dts-gen */

declare const decodeLib: {
    bitArray: {
        bitLength: any;
        bitSlice: any;
        byteswapM: any;
        clamp: any;
        concat: any;
        equal: any;
        extract: any;
        getPartial: any;
        partial: any;
    };
    cipher: {
        aes: any;
    };
    codec: {
        arrayBuffer: {
            fromBits: any;
            hexDumpBuffer: any;
            toBits: any;
        };
        base64: {
            fromBits: any;
            toBits: any;
        };
        base64url: {
            fromBits: any;
            toBits: any;
        };
        utf8String: {
            fromBits: any;
            toBits: any;
        };
    };
    exception: {
        bug: any;
        corrupt: any;
        invalid: any;
        notReady: any;
    };
    hash: {
    };
    keyexchange: {
    };
    misc: {
    };
    mode: {
        cbc: {
            decrypt: any;
            encrypt: any;
            name: string;
        };
    };
}

export const decodeLib